// gantryweb/Frontend/Gantry/src/app/common/models/event-feed-api-urls.model.ts
export class EventFeedApiUrls {
    eventsApi: string;
    racingContentEventsApi: string;
    nonRunnersApi: string;
    eventsResultApi: string;
    meetingResultsApi: string;
    runningOnTotalsApi: string;
    multiEventApi : string;
    latestResultsApi: string;
    avrVideoUrl:{
        [attr: string]: string;
    };
    avrApi: string;
    epsApi: string;
    signalRHubMessageApi: string;
    signalrRetryTimeOnDisconnect: number;
    snapShotDataTimeOut : number;
}